for fatima

+++ cart/views.py
@@ -11,19 +11,38 @@ import random
 
 def index(request):
     if not request.user.is_authenticated:
+        
+            context = {
+                "restaurants": Restaurant.objects.all(),
+            }
+            messages.warning(request, "Please login to view cart.", context)
+            return render(request, "cart/index.html")
+
+    
+    try:
+        cart = Cart.objects.get(user=request.user)
+        try:
+            orders = Order.objects.filter(user=request.user)
+            context = {
+                "user": request.user,
+                "cart_items": Cart_Item.objects.filter(cart=cart),
+                "restaurants": Restaurant.objects.all(),
+                'payment_choices': payment_choice,
+                "orders": orders
+            }
+        except Order.DoesNotExist:
+            context = {
+                "user": request.user,
+                "cart_items": Cart_Item.objects.filter(cart=cart),
+                "restaurants": Restaurant.objects.all(),
+                'payment_choices': payment_choice
+            }
+    except Cart.DoesNotExist:
         context = {
-            "restaurants": Restaurant.objects.all()
+            "user": request.user,            
+            "restaurants": Restaurant.objects.all(),
+            'payment_choices': payment_choice
         }
-        messages.warning(request, "Please login to view cart.", context)
-        return render(request, "cart/index.html")
-
-    cart = Cart.objects.get(user=request.user)
-
-    context = {
-        "user": request.user,
-        "cart_items": Cart_Item.objects.filter(cart=cart),
-        "restaurants": Restaurant.objects.all()
-    }
     return render(request, "cart/index.html", context)
 
 def cart_item(request, item):
@@ -69,6 +88,10 @@ def order(request):
 
         cart = Cart.objects.get(user = request.user)
         cart_item = Cart_Item.objects.filter(cart = cart)
+        username =  request.POST.get("username")
+        email = request.POST.get("email")
+        payment = request.POST.get("payment")
+        note = request.POST.get("note")
         address = request.POST.get("address")
         phone_number = request.POST.get("phone_number")
         total = request.POST.get("total")
@@ -81,9 +104,14 @@ def order(request):
         for item in cart_item:
             print(item.total)
             order_items = Order_Items.objects.create(product = item.product, quantity=item.quantity,total=item.total, cart = item.cart)
-            order = Order.objects.create(user=request.user ,order_items = order_items, address=address, phone_number=phone_number, total = total, deliverer=deliverer) 
-            cart_item = Cart_Item.objects.get(id = item.id)
-            cart_item.delete()
-
-        return JsonResponse({"success": True})
-    return JsonResponse({"Success":False})
+            if email != request.user:
+                order = Order.objects.create(user=request.user ,order_items = order_items, payment=payment, note=note ,address=address, phone_number=phone_number, total = total, deliverer=deliverer) 
+                cart_item = Cart_Item.objects.get(id = item.id)
+                cart_item.delete()
+            else:
+                messages.success(request,"Enter correct email.")
+                return render(request, "cart/index.html", context)
+        messages.success(request,"Order has been placed. It will be delivered soon.")
+        return render(request, "cart/index.html")
+    messages.warning(request, "Your cart is empty.")
+    return render(request, "cart/index.html")
